requirements-dev.txt
-r requirements.txt
black>=24.3.0
build>=0.7.0
isort>=5.13.2
twine>=3.8.0
types-setuptools>=69.2.0
setuptools>=69.2.0

# for test.py
python-dotenv>=1.0.1
